Add proper x translate
Based on bottom of column aligned based on roll angle. Does only suppoer
pitch that is increasing from left to right(folowing the clock) which is
the most natural. If you want something else offsets to accomodate
overlaps between columns must be added. That varialbe is not added yet
but is on to do list.